requirements.txt
pillow
pymupdf
opencv-python-headless
reportlab
streamlit==1.24.0
streamlit-drawable-canvas==0.9.2